Separate out flava modelas architecture
